Refactor SankeyVisualization to enhance functionality and user experience
- Introduced new components for improved visualization, including EnhancedNode, EnhancedLink, TagsToolbar, and OptionalMarkersOverlay.
- Streamlined tooltip management by integrating useSankeyTooltip for better interaction and data display.
- Updated zoom and pan controls with a dedicated ZoomControls component, enhancing navigation within the visualization.
- Removed deprecated code related to tag management and zoom functionality, simplifying the component structure.
- Improved data handling and layout tracking for better performance and responsiveness.
