+++ src/components/funnel/visualizations/funnel-graph/constants.ts
@@ -1,0 +1,25 @@
+// Shared constants for Sankey visualization
+
+export const STEP_COLORS = {
+  main: {
+    landing: '#3b82f6',
+    destination: '#10b981',
+    travelDetails: '#8b5cf6',
+    booking: '#f59e0b',
+    confirmation: '#ef4444',
+  },
+  split: {
+    domestic: '#34d399',
+    international: '#60a5fa',
+  },
+  links: {
+    main: '#93c5fd',
+    domestic: '#6ee7b7',
+    international: '#93c5fd',
+    high: '#10b981',
+    medium: '#f59e0b',
+    low: '#6b7280',
+  }
+} as const;
+
+
