+++ src/components/funnel/visualizations/funnel-graph/hooks/useEnhancedSankeyData.ts
@@ -1,0 +1,96 @@
+import React from 'react';
+
+type NodeInput = {
+  name: string;
+  value?: number;
+  color?: string;
+  index?: number;
+};
+
+type LinkInput = {
+  source: number;
+  target: number;
+  value: number;
+  sourceId?: string;
+  targetId?: string;
+  conversionRate?: number;
+  sourceValue?: number;
+};
+
+export function useEnhancedSankeyData(rechartsData: { nodes: NodeInput[]; links: LinkInput[] }) {
+  const enhancedData = React.useMemo(() => {
+    return {
+      nodes: rechartsData.nodes.map((node, index) => ({
+        name: node.name,
+        value: node.value || 0,
+        color: node.color,
+        index,
+        conversionRate: 0,
+      })),
+      links: rechartsData.links.map(link => ({
+        source: link.source,
+        target: link.target,
+        value: link.value || 0,
+        sourceId: link.sourceId || '',
+        targetId: link.targetId || '',
+        conversionRate: link.conversionRate || 0,
+        sourceValue: link.sourceValue || 0,
+      })),
+    };
+  }, [rechartsData]);
+
+  const isValidData = React.useMemo(() => {
+    return (
+      enhancedData.nodes.length > 0 &&
+      enhancedData.links.length > 0 &&
+      enhancedData.nodes.every(n => typeof n.value === 'number' && n.value > 0) &&
+      enhancedData.links.every(l => typeof l.value === 'number' && l.value > 0) &&
+      enhancedData.links.every(l => l.source >= 0 && l.target >= 0 && l.source < enhancedData.nodes.length && l.target < enhancedData.nodes.length)
+    );
+  }, [enhancedData]);
+
+  const preCalculatedStepHeights = React.useMemo(() => {
+    const heights: Record<string, number> = {};
+    const maxValue = Math.max(...enhancedData.nodes.map(n => n.value || 0));
+    const minHeight = 20;
+    const maxHeight = 170;
+    enhancedData.nodes.forEach((node, index) => {
+      const nodeValue = node.value || 0;
+      let calculatedHeight = minHeight;
+      if (maxValue > 0) {
+        calculatedHeight = Math.max(minHeight, Math.min(maxHeight, (nodeValue / maxValue) * 120 + 20));
+      }
+      heights[node.name] = calculatedHeight;
+      heights[`step-${index}`] = calculatedHeight;
+    });
+    return heights;
+  }, [enhancedData.nodes]);
+
+  const dynamicContainerHeight = React.useMemo(() => {
+    const splitNodes = enhancedData.nodes.filter(node => node.name.includes('split'));
+    const maxSplits = Math.max(
+      ...enhancedData.nodes.map(node => {
+        const stepName = node.name;
+        const stepIndex = enhancedData.nodes.findIndex(n => n.name === stepName);
+        const nextStep = enhancedData.nodes[stepIndex + 1];
+        if (nextStep && nextStep.name.includes('split')) {
+          return enhancedData.nodes.filter(n => n.name.includes('split') && n.name.startsWith(stepName.split('-')[0])).length;
+        }
+        return 0;
+      })
+    );
+    const baseHeight = 440;
+    const extraHeightPerSplit = 40;
+    const totalExtraHeight = Math.max(0, (isFinite(maxSplits) ? maxSplits : 0) * extraHeightPerSplit);
+    return baseHeight + totalExtraHeight;
+  }, [enhancedData.nodes]);
+
+  return {
+    enhancedData,
+    isValidData,
+    preCalculatedStepHeights,
+    dynamicContainerHeight,
+  } as const;
+}
+
+
